Created Movement System
The only time it doesn't work is if the player is moving while the board is rotating. HOWEVER, during the game, the board only rotates between turns, meaning that the player would not be allowed to move while the board is rotating. This will be fixed once turns are implemented.